LayoutTests/canvas/philip/tests/2d.gradient.radial.negative.html

<!DOCTYPE html>
<title>Canvas test: 2d.gradient.radial.negative</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="2d.gradient.linear.transform.3.html" title="2d.gradient.linear.transform.3">
<link rel="next" href="2d.gradient.radial.nonfinite.html" title="2d.gradient.radial.nonfinite">
<body class="show_output">
<p>
 <a href="2d.gradient.linear.transform.3.html" accesskey="p" title="[p] 2d.gradient.linear.transform.3">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="2d.gradient.radial.nonfinite.html" accesskey="n" title="[n] 2d.gradient.radial.nonfinite">&gt;</a>
<h1><a href="index.2d.html">2d</a>.<a href="index.2d.gradient.html">gradient</a>.<a href="index.2d.gradient.radial.html">radial</a>.negative</h1>
<p class="desc">createRadialGradient() throws INDEX_SIZE_ERR if either radius is negative</p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.2d.gradient.radial.negative">2d.gradient.radial.negative</a>

</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

try { var _thrown = false;
  ctx.createRadialGradient(0, 0, -0.1, 0, 0, 1);
} catch (e) { if (e.code != DOMException.INDEX_SIZE_ERR) _fail("Failed assertion: expected exception of type INDEX_SIZE_ERR, got: "+e.message); _thrown = true; } finally { _assert(_thrown, "should throw exception of type INDEX_SIZE_ERR: ctx.createRadialGradient(0, 0, -0.1, 0, 0, 1)"); }
try { var _thrown = false;
  ctx.createRadialGradient(0, 0, 1, 0, 0, -0.1);
} catch (e) { if (e.code != DOMException.INDEX_SIZE_ERR) _fail("Failed assertion: expected exception of type INDEX_SIZE_ERR, got: "+e.message); _thrown = true; } finally { _assert(_thrown, "should throw exception of type INDEX_SIZE_ERR: ctx.createRadialGradient(0, 0, 1, 0, 0, -0.1)"); }
try { var _thrown = false;
  ctx.createRadialGradient(0, 0, -0.1, 0, 0, -0.1);
} catch (e) { if (e.code != DOMException.INDEX_SIZE_ERR) _fail("Failed assertion: expected exception of type INDEX_SIZE_ERR, got: "+e.message); _thrown = true; } finally { _assert(_thrown, "should throw exception of type INDEX_SIZE_ERR: ctx.createRadialGradient(0, 0, -0.1, 0, 0, -0.1)"); }


});
</script>
